+++ src/test/java/com/uwetrottmann/trakt5/services/SyncTest.java
@@ -52,6 +52,7 @@ import com.uwetrottmann.trakt5.enums.ExtendedShowsWatched;
 import com.uwetrottmann.trakt5.enums.HistoryType;
 import com.uwetrottmann.trakt5.enums.Rating;
 import com.uwetrottmann.trakt5.enums.RatingsFilter;
+import com.uwetrottmann.trakt5.enums.Specials;
 import org.junit.Test;
 import org.threeten.bp.Instant;
 import org.threeten.bp.OffsetDateTime;
@@ -357,7 +358,7 @@ public class SyncTest extends BaseTestCase {
     @Test
     public void test_watchedShows() throws IOException {
         Response<List<BaseShow>> response = executeCallWithoutReadingBody(
-                getTrakt().sync().watchedShows(PAGE_ONE, LIMIT_MAX, ExtendedShowsWatched.PROGRESS));
+                getTrakt().sync().watchedShows(PAGE_ONE, LIMIT_MAX, ExtendedShowsWatched.PROGRESS, Specials.TRUE));
 
         assertListPaginationHeaders(response);
         assertSyncShows(response.body(), "watched");
@@ -368,7 +369,8 @@ public class SyncTest extends BaseTestCase {
         // Note: Starting 2026-05-30 these extended values will be the default
         List<BaseShow> shows = executeCall(
                 getTrakt().sync().watchedShows(PAGE_ONE, LIMIT_MAX,
-                        ExtendedShowsWatched.of(ExtendedShowsWatched.FULL, ExtendedShowsWatched.NOSEASONS)));
+                        ExtendedShowsWatched.of(ExtendedShowsWatched.FULL, ExtendedShowsWatched.NOSEASONS),
+                        Specials.TRUE));
 
         assertWatchedShowsFullNoSeasons(shows);
     }
@@ -376,7 +378,7 @@ public class SyncTest extends BaseTestCase {
     @Test
     public void test_watchedShowsMin() throws IOException {
         Response<Map<String, Map<String, Map<String, List<OffsetDateTime>>>>> response = executeCallWithoutReadingBody(
-                getTrakt().sync().watchedShowsMin(PAGE_ONE, LIMIT_MAX));
+                getTrakt().sync().watchedShowsMin(PAGE_ONE, LIMIT_MAX, Specials.TRUE));
 
         // As of 2026-05-08, pagination is not supported, yet
         // assertListPaginationHeaders(response);
@@ -387,7 +389,7 @@ public class SyncTest extends BaseTestCase {
     @Test
     public void test_watchedEpisodes() throws IOException {
         Response<List<WatchedEpisode>> response = executeCallWithoutReadingBody(
-                getTrakt().sync().watchedEpisodes(PAGE_ONE, LIMIT_MAX));
+                getTrakt().sync().watchedEpisodes(PAGE_ONE, LIMIT_MAX, Specials.TRUE));
 
         assertListPaginationHeaders(response);
         assertThat(response.body())
@@ -401,7 +403,7 @@ public class SyncTest extends BaseTestCase {
     @Test
     public void test_watchedEpisodesMin() throws IOException {
         Response<Map<String, List<OffsetDateTime>>> response = executeCallWithoutReadingBody(
-                getTrakt().sync().watchedEpisodesMin(PAGE_ONE, LIMIT_MAX));
+                getTrakt().sync().watchedEpisodesMin(PAGE_ONE, LIMIT_MAX, Specials.TRUE));
 
         // As of 2026-05-08, pagination is not supported, yet
         // assertListPaginationHeaders(response);